{{ cookiecutter.project_slug }}/requirements.txt
aiohttp==3.6.2
aiohttp-devtools==0.13.1
aiohttp-swagger==1.0.12
environs==7.1.0
ipdb==0.12.3
isort==4.3.21
pytest==5.3.1
pytest-aiohttp==0.3.0
pytest-cov==2.8.1
PyYAML==5.2
sentry-sdk==0.14.0
